Move more code outside of main. Created Win32Midi module.
